Laid groundwork for langevin ID computation and made changes to recompute Inpaint metrics with current setup

+++ fjd_metric.py
@@ -464,7 +464,7 @@ class FJDMetric:
 
 def get_embedding_statistics(embeddings, cuda=False):
     if cuda:
-        embeddings = embeddings.double()  # More precision = more stable
+        embeddings = embeddings  # More precision = more stable
         mu = torch.mean(embeddings, 0)
         sigma = torch_cov(embeddings, rowvar=False)
     else:
